_deprecated_backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# AI and ML
openai>=1.0.0
elevenlabs>=0.2.0

# Data processing
pandas>=1.5.0

# Database
supabase>=2.0.0

# Authentication (optional)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File handling
aiofiles>=23.1.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0